+++ DynatracePS/Public/Get-DynatraceGroupPermission.ps1
@@ -19,10 +19,13 @@ function Get-DynatraceGroupPermission {
         Output the properties as a JSON string
         
     .EXAMPLE
-        Get-DynatraceGroupPermission -groupUuid e5e9b12d-daf8-40d0-a6b5-7094667dd142
+        Get-DynatraceGroupPermission -groupUuid e5e9b12d-daf8-40d0-a6b5-7094667dd142 -OutputNotNested
 
     .EXAMPLE
         Get-DynatraceGroupPermission -GroupName 'Monitoring Viewer'
+    
+    .EXAMPLE
+        Get-DynatraceGroup | %{Get-DynatraceGroupPermission -groupUuid $_.uuid}
 
     .NOTES
         https://api.dynatrace.com/spec/#/Group%20management/GroupsController_getUsersForGroup
@@ -62,7 +65,20 @@ function Get-DynatraceGroupPermission {
             $result = Invoke-DynatraceAccountManagementAPIMethod @splatParameters
 
             if ($OutputNotNested) {
+                # caching management zones for lookup
+                #
+                $mz = Get-DynatraceManagementZone
                 foreach ($permission in $result.permissions) {
+                    if ($permission.scope -match ':') {
+                        # if there is colon in permissions scope, it will be a string
+                        # with environment name:management zone id
+                        #
+                        $splitPermissionScope = $permission.scope -split ':'
+                        $mzName = ($mz | Where-Object{$_.ID -eq $splitPermissionScope[1]}).Name
+                        $scope = "{0}:{1}" -f $splitPermissionScope[0],$mzName
+                    } else {
+                        $scope = $permission.scope;
+                    }
                     $output += [pscustomobject]@{
                         uuid = $result.uuid;
                         name = $result.name;
@@ -72,7 +88,7 @@ function Get-DynatraceGroupPermission {
                         createdAt = $result.createdAt;
                         updatedAt = $result.updatedAt;
                         permissionName = $permission.permissionName;
-                        Scope = $permission.scope;
+                        Scope = $scope;
                         ScopeType = $permission.scopeType;
                         permissionCreatedAt = $permission.createdAt;
                         permissionUpdatedAt = $permission.updatedAt;