+++ Changelog.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.12]
+### Added
+-Get-DyntraceUserLastLogin to get user name and last login time
+### Modified
+-Get-DynatraceGroupPermission to output MZ name instead of MZ id so it's easier to grok
+
 ## [0.1.11]
 ### Modified
 -Replaced Invoke-RestMethod with Invoke-WebRequest in order to provide better support across various PS versions